admin layers finished

+++ src/main/java/com/mmdev/dictionaryy/service/AdminService.java
@@ -6,11 +6,12 @@ import com.mmdev.dictionaryy.mapper.admin.AdminDtoMapper;
 import com.mmdev.dictionaryy.mapper.admin.AdminMapper;
 import com.mmdev.dictionaryy.model.AdminDto;
 import com.mmdev.dictionaryy.repository.AdminRepository;
+import jakarta.persistence.EntityExistsException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,9 +29,10 @@ public class AdminService {
 				.collect(Collectors.toList());
 	}
 
-	public Optional<AdminDto> getAdminById(Long id) {
+	public AdminDto getAdminById(Long id) {
 		return adminRepository.findById(id)
-				.map(adminDtoMapper::map);
+				.map(adminDtoMapper::map)
+				.orElseThrow(() -> new EntityNotFoundException("Admin not found with id: " + id));
 	}
 
 	public AdminDto createAdmin(AdminDto adminDto) {
@@ -43,15 +45,44 @@ public class AdminService {
 		Admin admin = adminRepository.findById(id)
 				.orElseThrow(() -> new EntityNotFoundException("Admin not found with id: " + id));
 
+		if (!admin.getEmail().equals(adminDto.email()) && adminRepository.existsByEmail(adminDto.email())) {
+			throw new EntityExistsException("Email is already in use by another admin.");
+		}
+
 		admin.setName(adminDto.name());
 		admin.setEmail(adminDto.email());
 		admin.setPassword(adminDto.password());
 
-		adminRepository.save(admin);
-		return adminDtoMapper.map(admin);
+		Admin savedAdmin = adminRepository.save(admin);
+		return adminDtoMapper.map(savedAdmin);
 	}
 
 	public void deleteAdminById(Long id) {
+		if(adminRepository.findById(id).isEmpty()){
+			throw new EntityNotFoundException("Admin not found with id: " + id);
+		}
 		adminRepository.deleteById(id);
+	}
+
+	public AdminDto patchAdminById(Long id, Map<String, Object> updates) {
+		Admin admin = adminRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("Admin not found with id: " + id));
+
+		updates.forEach((key, value) -> {
+			switch (key) {
+				case "name":
+					admin.setName((String) value);
+					break;
+				case "email":
+					admin.setEmail((String) value);
+					break;
+				case "password":
+					admin.setPassword((String) value);
+					break;
+			}
+		});
+
+		Admin savedAdmin = adminRepository.save(admin);
+		return adminDtoMapper.map(savedAdmin);
 	}
 }
